Compile value converter expressions directly into ValueBuffer factory
This moves converting a store value to happen immediately after the value is read from the data reader. This means that it only needs to happen in one place, which will make it cleaner when implementing more complex conversions; such as those needed for Oracle.

Value converters are now registered as expressions, and the expression is compiled directly into the materializer code with no lambda or method invocation, which should be optimal for perf.